requirements.txt
beautifulsoup4==4.7.1
python-telegram-bot==11.1.0
requests==2.22.0
urllib3==1.25.3
psycopg2==2.8.3 